헤더파일 이름과 cpp 파일 이름 통일화
헤더와 cpp파일 이름 맞춰서 커밋

+++ Basic_B_1/Character.cpp
@@ -9,7 +9,7 @@ ACharacter::ACharacter(string NewName, int NewHp, int NewAtk, int NewDef, int Ne
     Atk = NewAtk;
     Def = NewDef;
     
-    Critical = (int)NewCri; 
+    Critical = NewCri; 
 
     cout << "Character 생성됨: " << Name << " (HP: " << Hp << ", ATK: " << Atk << ")" << endl;
 }
